improve sam model
